Fix TypeError in mark_messsages_as_read: add default user from session
Previously, the function mark_messsages_as_read was defined as:
    def mark_messsages_as_read(user, channel=None, parent_channel=None):

This caused a TypeError ("missing 1 required positional argument: 'user'") when it was called without the user parameter.

To fix this, the function is modified to set the user to frappe.session.user if no user is provided. This ensures that the function always receives a valid user. For example, the updated function signature is:

    def mark_messsages_as_read(user=None, channel=None, parent_channel=None):
        if not user:
            user = frappe.session.user
        ...

This change resolves the error and improves the robustness of the API.